requirements.txt
Babel==2.2.0
Django==1.9.2
Fabric==1.10.2
Markdown==2.6.5
Pillow==3.1.1
PyJWT==1.4.0
PyYAML==3.11
South==1.0.2
Unidecode==0.04.19
argparse==1.2.1
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
certifi==2016.2.28
chance==0.102
django-classy-tags==0.7.1
django-countries==3.4.1
django-filter==0.12.0
django-phonenumber-field==1.0.0
django-polymorphic==0.9.1
django-post-office==2.0.6
django-rest-swagger==0.3.5
django-sekizai==0.9.0
django-shop==0.2.0
django-versatileimagefield==1.3
djangorestframework==3.3.2
ecdsa==0.13
httplib2==0.9.2
jsonfield==1.0.3
oauthlib==1.0.3
paramiko==1.16.0
phonenumbers==7.2.5
psycopg2==2.6.1
pycrypto==2.6.1
python-magic==0.4.11
python-openid==2.2.5
python-slugify==1.2.0
python-social-auth==0.2.14
pytz==2015.7
redis==2.10.5
requests==2.9.1
requests-oauthlib==0.6.1
selenium==2.52.0
singledispatch==3.4.0.3
six==1.10.0
sockjs-tornado==1.0.2
tornado==4.3
twilio==5.4.0
